Edit: Modularize Code
