+++ src/modules/counter.js
@@ -1,0 +1,7 @@
+const mealCounter = () => {
+  const container = document.querySelector('.container');
+  const mealcounter = document.querySelector('.meal-count');
+  mealcounter.innerHTML = `(${container.children.length} recipes available)`;
+};
+
+export default mealCounter;
